fix: normalize waste reminder channel configs
