+++ packages/auth-runtime/src/waste-management/core/operations.ts
@@ -6,8 +6,8 @@ import { resolveActorInfo } from '../../iam-account-management/shared.js';
 import { validateCsrf } from '../../shared/request-security.js';
 import { asApiItem, createApiError, parseRequestBody, requireIdempotencyKey } from '../../shared/request-helpers.js';
 import { authorizeWasteManagementAction, emitWasteAuditEvent, getAuthorizedWasteManagementInstanceId } from './auth.js';
+import { wasteManagementOperationSchemas } from './operation-schemas.js';
 import { startPluginOperationJobFromFacade } from './operations-support.js';
-import { wasteManagementOperationSchemas } from './schemas.js';
 import { loadConfiguredWasteSettings } from './settings-shared.js';
 import type { WasteManagementHandlerDeps } from './types.js';
 import { getRequestId } from './utils.js';
