+++ packages/core/src/waste-management-master-data.ts
@@ -1,4 +1,5 @@
 export * from './waste-management/master-data-contract.js';
+export * from './waste-management/master-data-localized-text.js';
 export * from './waste-management/master-data-addresses.js';
 export * from './waste-management/master-data-tours.js';
 export * from './waste-management/master-data-scheduling.js';
