+++ packages/core/src/waste-management/master-data-localized-text.ts
@@ -1,0 +1,1 @@
+export type WasteLocalizedTextRecord = Readonly<Record<string, string>>;
